New feature: M746 to activate/deactivate/status of external probe connected to secure_sw ========================================================================================
M746 - setting of an external endstop to be used as an external endstop zprobe.
M746 S1 to enable this functionality.
M746 S0 to disable this functionality.
M746 to see the status of this functionality.

The actual pin is provided by EXTERNAL_ENDSTOP_Z_PROBING_PIN macro, which defaults to pin 71 (in totumduino's silk screen "Secure_sw", see also M743).

If enabled you need to keep an external zprobe connected to this totumduino connector that makes the endstop read "open" in normal state or
you will have undesirable behaviour (only during the time a Z probe is done, so during homing, G30 and G38). It is a good design to execute
M746 S1 before a G38 and M746 S0 afterwards.

A possible probe is an electrical continuity probe between a copper cad (for making PCBs) and the mill (that in the hybrid head is to GND), like this:

 |  | secure_sw
 |  \------------------------------+----- attach this to copper clad (via a metal washer to which the wire is soldered?)
 \-----------------------/\/\/\/---|
                         1 kOhm