Feature improvement: Variable feedrate for G38 ==============================================
G38 now may take a parameter like G38 S100 to override the feedrate to be used during probing.

It introduces a safety measure that under no circumstances the parameter is set to a value over 200, which is considered to be the highest safe value for probing.

Higher values tend to introduce inaccuracy in the read value. High values with pointy mills (like 0.1 mm V shaped 10 degrees mills) may lead to the point punching the PCB clad.