New Functionality: Notify FAB-UI of impossible endstop combinations for extended functionality
This includes two new error types to notify FABUI when:
	1) Both Y Endstops are pressed at the same time
	2) Both Z Endstops are pressed at the same time

The underlying idea is to allow these combinations to be used to trigger some functionality, like for example shutdown.

It is to be noted that this does not require that actually both endstops are simulatenously triggered as pointed out by Randy,
some hardware modification (wiring) can be put in place, so that when an new hardware switch is activated, the totumduino perceives
this as both endstop having been activated.