Fixed External Endstop Port logic status inverted.
External Z probing has been fixed. to enable: M746 S1 when the external
probe is inserted in the secure_sw port
M746 S0 when the external probe is disconnected

Do not use the external probe without enabling the feature.
Do not enable the feature if the external probe is not connected.
The M746 S1 and S0 comands are not saved in eeprom, so those are lost
upon resetting the controller or rebooting the FABtotum.